Update MPI handling to automatically setup MPI compilers when enable_mpi is ON
Instead of checking to make sure that the compiler given is capable of compiling
MPI actually do a search for MPI and tie up all the include paths, library paths,
library linking, etc.

Currently the properties are all set for the pebbl library which everything
depends on so all of the dependencies are automatically tied up. If more
libraries are created later they will need this setup as well.