Fix issue when MPI compiler has compile flags
Fixes issue #4

The call to target_compile_options which was only called for the pebbl library when
cmake determined that mpi has compile options was missing a necessary parameter that
determines the transitivity of the options. I set this to public to ensure that all
libraries or executables that depend on pebbl will also get any options from mpi.

I also updated the target_link_libraries call to explicitly state that the mpi
libraries are also public for the same reason as above.